refactor: header 생성 책임 팩토리로 분리
